Rename "reverse_existence" checks to "restricted"
Also, factor out word lists into module-level variables, for
access from elsewhere.
