+++ proselint/checks/restricted/__init__.py
@@ -1,0 +1,1 @@
+"""Restricted word lists."""
