--- proselint/checks/reverse_existence/__init__.py
@@ -1,1 +1,0 @@
-"""Reverse existence."""
