+++ tests/test_restricted_top1000.py
@@ -1,12 +1,12 @@
 """Tests reverse existence for top1000 and check."""
 
-from proselint.checks.reverse_existence import top1000 as chk
+from proselint.checks.restricted import top1000 as chk
 
 from .check import Check
 
 
 class TestCheck(Check):
-    """The test class for reverse_existence.top1000."""
+    """The test class for restricted.top1000."""
 
     __test__ = True
 
@@ -16,7 +16,7 @@ class TestCheck(Check):
         return chk
 
     def test_smoke(self):
-        """Basic smoke test for reverse_existence.top1000."""
+        """Basic smoke test for restricted.top1000."""
         assert self.passes("""I am blonde.""")
         assert self.passes("""I'm gonna listen to music tonight.""")
         assert self.passes("""I will go to sleep because I have school.""")
